Pick the GPG key with the lowest TTL.
The assumption is that there can be keys in the local keyring that
are not yet active, because they have just been generated but have
not been included into the archive's keyring yet. So pick the key
with the lowest expiry that is still valid instead. This prefers
current keys over keys just generated.

Once keys are removed from the archive, they then also need to be
pruned from the local keyring (which is a good thing for other
reasons as well).

Fixes #6