Log dupload failures (#7)
retrying in stretch does not support passing in logger=logging, so
we need to do explicit logging of dupload failures - even though you
will also have the literal output of dupload on stderr/stdout it
will then proceed to sleep between retries and the exception hopefully
makes this clearer.

Also adjust the unit of the sleep time, which is actually in
milliseconds instead of seconds.