Clear up the remaining mypy confusion.
Declare all optional instance variables explicitly and with type hints.
This works both on stretch and sid. Fix up the GPG key assigment to
access bits of the output independently rather than using a tuple
assignment. Run mypy on Gitlab CI to ensure that we do not accidentally
break typing.