Pick the maintainer email passed to sbuild from the key's UID.
In the Debian setup we need to put a valid email address into the
.changes file because dak is mailing us about the uploads. On the
other hand the set of configured emails matches a very strict
pattern that is used for buildd autosigning key generation and
checked by a pattern within dak's key acceptance scripts.

So this change fetches the email address from the key's UID
instead of constructing it from the architecture and hostname
of the buildd, which does not match in case a buildd is building
for multiple architecture, which is the case for an amd64 buildd
that also builds for i386.

As part of this refactor the key picking logic to return an
object to ensure type safety and ensure that the right field
value ends up in the sbuild invocation.

Fixes #10