Make sure the main buildd process is the only one to get SIGTERM (#12)
The current plan would be to store the received SIGTERM and ask sbuild
to terminate gracefully. This might be revisited but the current default
of control-group does not actually work as an unprivileged user.

The documented behavior of KillMode=mixed implies that only the main
process gets SIGTERM, but if it does not react in time all processes
spawned by the service will be reaped using SIGKILL. That is sort of
what we want, except that sbuild does not get to cleanup in this case.

So a follow-up change needs to make sure that sbuild gets the signal.