buildd.service: changes Requires into Wants (#15)
When the SSH connection to wanna-build dies (e.g. network hiccup,
buildd.debian.org reboot, etc.) the buildd-ssh-master.service is
restarted. Given that buildd.service Requires the former, it also
gets restarted, killing the build.

Replace the Requires into a Wants as a control master is not strictly
needed to run buildd.service. In the worst case a new socket is created,
but the previous commit ensures it get removed the next time
buildd-ssh-master.service is (re)started.