Fix version when calling wanna-build --uploaded
wanna-build is not really consistent in the way the version should be
passed. It sometimes require the source version, sometimes the binary
version (including a possibly +bX):
 --take only accepts the source version
 --attempted and --built accept both source and binary version
 --give-back accepts both source and binary version
 --uploaded only accepts the binary version

Currently pybuildd only passes the source version, which causes binNMUed
packages to never go through the "uploaded" state. This patch fixes that
by using the binary version wherever possible (i.e. for --attempted,
--built, --give-back and --uploaded) as this prevents a change state for
the wrong version.